requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
plotly>=5.20.0
anthropic>=0.40.0
yfinance>=0.2.40
python-dotenv>=1.0.0
